add rake task to fetch movies for a day

+++ db/migrate/20171125010244_create_box_office_days.rb
@@ -6,7 +6,7 @@ class CreateBoxOfficeDays < ActiveRecord::Migration[5.1]
 
             t.integer :metacritic_score
 
-            t.float :imdb_rating
+            t.integer :imdb_rating
             t.integer :imdb_vote_count
 
             t.string :tomato_consensus
